abstracting class
