+++ app/src/main/java/ca/klapstein/baudit/presenters/PatientHomePresenter.java
@@ -1,19 +1,13 @@
 package ca.klapstein.baudit.presenters;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
 import android.util.Log;
 import ca.klapstein.baudit.data.Account;
-import ca.klapstein.baudit.data.Patient;
 import ca.klapstein.baudit.data.Problem;
 import ca.klapstein.baudit.data.ProblemTreeSet;
 import ca.klapstein.baudit.views.HomeView;
 import ca.klapstein.baudit.views.PatientHomeView;
 import ca.klapstein.baudit.views.ProblemRowView;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 
 /**
  * MVP presenter for presenting a {@code Patient}'s home screen on a {@code HomeView}.
@@ -22,12 +16,9 @@ import java.util.Collections;
  * @see Account
  * @see ProblemTreeSet
  */
-public class PatientHomePresenter extends Presenter<PatientHomeView> {
+public class PatientHomePresenter extends ProblemListPresenter<PatientHomeView> {
 
     private static final String TAG = "PatientHomePresenter";
-    private Patient patient;
-    @NonNull
-    private ProblemTreeSet problemTreeSet = new ProblemTreeSet();
 
     public PatientHomePresenter(PatientHomeView view, Context context) {
         super(view, context);
@@ -99,35 +90,5 @@ public class PatientHomePresenter extends Presenter<PatientHomeView> {
             Log.e(TAG, "failed to delete problem", e);
             view.updateDeleteProblemError();
         }
-    }
-
-    public int getTrueProblemIndex(int position) {
-        try {
-            Problem problem = (Problem) problemTreeSet.toArray()[position];
-            return new ArrayList<Problem>(Arrays.asList(patient.getProblemTreeSet().toArray(new Problem[0]))).indexOf(problem);
-        } catch (Exception e) {
-            Log.e(TAG, "failed to obtain true problem index", e);
-            view.updateAccountLoadError();
-            return 0;
-        }
-    }
-
-    public void filterProblemsByKeyWords(CharSequence constraint) {
-        try {
-            problemTreeSet.clear();
-            problemTreeSet.addAll(patient.getProblemTreeSet());
-        } catch (Exception e) {
-            Log.e(TAG, "failed to obtain patient account info", e);
-        }
-
-        ArrayList<String> searchTokens = new ArrayList<>(Arrays.asList(constraint.toString().toLowerCase().split(" ")));
-        Log.d(TAG, "filtering with tokens: " + searchTokens);
-        Problem[] problemArray = problemTreeSet.toArray(new Problem[0]);
-        for (Problem aProblemArray : problemArray) {
-            if (Collections.disjoint(searchTokens, aProblemArray.getKeywords()))
-                problemTreeSet.remove(aProblemArray);
-        }
-        view.updateList();
-        view.updateProblemNumber(problemTreeSet.size());
     }
 }
