+++ app/src/main/java/ca/klapstein/baudit/views/CareProviderProblemListView.java
@@ -1,6 +1,6 @@
 package ca.klapstein.baudit.views;
 
-public interface CareProviderProblemListView extends View {
+public interface CareProviderProblemListView extends ProblemListView {
     void updateList();
     void updateAccountLoadError();
 }
