+++ app/src/main/java/ca/klapstein/baudit/views/PatientHomeView.java
@@ -1,7 +1,12 @@
 package ca.klapstein.baudit.views;
 
-public interface PatientHomeView extends HomeView {
+public interface PatientHomeView extends ProblemListView {
     void updateDeleteProblemError();
 
+    void updateUsernameDisplay(String username);
+
+    void updateEmailDisplay(String email);
+
+    void updateAccountLoadError();
     void updateProblemNumber(int problemNumber);
 }
