+++ app/src/main/java/ca/klapstein/baudit/views/ProblemListView.java
@@ -1,0 +1,9 @@
+package ca.klapstein.baudit.views;
+
+public interface ProblemListView extends View {
+    void updateAccountLoadError();
+
+    void updateList();
+
+    void updateProblemNumber(int problemNumber);
+}
